[Refactor] 코드 파일 분리
